moved routes outside of utils
