+++ didact/src/Routes.jsx
@@ -1,11 +1,11 @@
 import React from "react";
 import { Route} from "react-router-dom";
-import PrivateRoute from "./PrivateRoute";
-import FormikLoginForm from "../components/login/Login.jsx";
-import FormikRegisterForm from "../components/register/Register.jsx";
+import PrivateRoute from "./utils/PrivateRoute";
+import FormikLoginForm from "./components/login/Login.jsx";
+import FormikRegisterForm from "./components/register/Register.jsx";
 // import Dashboard from "../components/dashboard/dashboard.jsx";
-import Auth from '../auth/Auth';
-import MainPage from '../components/mainPage/MainPage'
+import Auth from './auth/Auth';
+import MainPage from './components/mainPage/MainPage'
 
 const Routes = () => {
     return (
